style(.editorconfig): 코드 스타일 파일 설정
